Fix bug in providers schema - address sub-type
This change fixes a bug in the definition of the providers schema that was
causing false-positives during validation of provider JSON files. In other
words, provider JSON files that *should* be flagged as invalid were passing
through the validator as valid.

The bug was that, in the syntax of the JSON schema file, the `address` sub-type
of the `addresses` array was not tagged with `"type": "object"`. This bit of
text lets the validator know that it should treat items of the array as objects
and to test them against the set of properties given. So while the set of
properties for a valid address was already defined, the lack of `"type":
"object"` meant that the validator essentially was just looking for any valid
JSON data type as elements of the array.